pyproject.toml
[project]
name = "rbac-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "rbac_service.settings"

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "single"
